backend/requirements.txt
flask
newsapi-python
textblob